Move isProtectedSource to import-requirements
